Adds field removal to ContentTypeStep (#8282)
